added groundwork for multiple pokemon variations
This version added the groundwork to be able to display in the pokedex
which pokemon have multiple variations. Stopped because it’s later at
night, I’m tired and also there’s a lot of work that needs to be done
on the data perspective in regards to Mega pokemon so I don’t have to
write multiple if statements.